backend/requirements.txt
fastapi
uvicorn
sentence-transformers
chromadb
pypdf2
python-docx
python-multipart
pymongo

